Adding basic Flask project to the back end
Created a new back-end folder. this folder along with front-end will be deleted once we integrate the two.

This is a project outline of a default Flask package in pycharm. Will include SQLAlchemy later this is a very rough outline. To follow the usage guide visit this link.

https://flask.palletsprojects.com/en/3.0.x/installation/

https://flask.palletsprojects.com/en/3.0.x/quickstart/

I would not pull this quite yet unless you are okay with a manual install. otherwise wait until I write a requirements.txt for the project